Refactor hasTwoColumnLayout function into separate utility file
- Move hasTwoColumnLayout function from layout.js to new utils.tsx file
- Update import paths in related files
- Improve type safety by adding TypeScript type annotations
